# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "以实现无纸化办公为目标",
        "title": "进销存系统",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/jxc",
    "paths": {
        "/cargo": {
            "put": {
                "description": "修改货品",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "修改货品",
                "parameters": [
                    {
                        "description": "货品和属性 和计量单位",
                        "name": "ReqAddCargo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ReqAddCargo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status 200 表示成功 否则提示msg内容",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "此接口用于新增原材料或制品， 新增时包括 计量和属性值",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "新增货品",
                "parameters": [
                    {
                        "description": "货品和属性 和计量单位",
                        "name": "ReqAddCargo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ReqAddCargo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status 200 表示成功 否则提示msg内容",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/cargo/_search": {
            "get": {
                "description": "通过货品code 或者 货品名称搜索 货品 (可以是原材料 和制品)",
                "summary": "搜索货品",
                "parameters": [
                    {
                        "type": "string",
                        "description": "货品code 或者 货品名称",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "类型 1:原材料 2:半成品 3:成品",
                        "name": "type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "页数",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "每页条数",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status 200 表示成功 否则提示msg内容",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.CargoKind"
                            }
                        }
                    }
                }
            }
        },
        "/cargo/{cargo_id}": {
            "get": {
                "description": "获取货品详情",
                "summary": "获取货品详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "货品ID",
                        "name": "cargo_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status 200 表示成功 否则提示msg内容",
                        "schema": {
                            "$ref": "#/definitions/dto.ReqAddCargo"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除货品",
                "summary": "删除货品",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "货品ID",
                        "name": "cargo_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status 200 表示成功 否则提示msg内容",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/cargo_kind": {
            "put": {
                "description": "修改货品种类",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "修改货品种类",
                "parameters": [
                    {
                        "description": "货品种类和属性",
                        "name": "货品种类",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CargoKindWithAttrs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status 200 表示成功 否则提示msg内容",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "新增货品种类，新增种类时同时新增种类相关规格属性",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "新增货品种类",
                "parameters": [
                    {
                        "description": "货品种类和属性",
                        "name": "货品种类",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CargoKindWithAttrs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status 200 表示成功 否则提示msg内容",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/cargo_kind/_search": {
            "get": {
                "description": "通过货品code 或者 货品名称搜索 货品种类",
                "summary": "搜索货品种类",
                "parameters": [
                    {
                        "type": "string",
                        "description": "货品code 或者 货品名称",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "页数",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "每页条数",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status 200 表示成功 否则提示msg内容",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.CargoKind"
                            }
                        }
                    }
                }
            }
        },
        "/cargo_kind/{ck_id}": {
            "get": {
                "description": "获取货品详情，获取货品详情和相关属性",
                "summary": "获取货品详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "货品种类ID",
                        "name": "ck_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status 200 表示成功 否则提示msg内容",
                        "schema": {
                            "$ref": "#/definitions/dto.CargoKindWithAttrs"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除货品",
                "summary": "删除货品",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "货品种类ID",
                        "name": "ck_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status 200 表示成功 否则提示msg内容",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/cargo_process": {
            "put": {
                "description": "修改制品生产流程",
                "summary": "修改制品生产流程",
                "parameters": [
                    {
                        "description": "文件",
                        "name": "processes",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.CargoProcess"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status 200 表示成功 否则提示msg内容",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "新增制品生产流程",
                "summary": "新增制品生产流程",
                "parameters": [
                    {
                        "description": "文件",
                        "name": "processes",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.CargoProcess"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status 200 表示成功 否则提示msg内容",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/cargo_process/{cargo_id}": {
            "put": {
                "description": "通过cargo_id制品生产流程",
                "summary": "获取制品生产流程",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "货品ID",
                        "name": "cargo_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status 200 表示成功 否则提示msg内容",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.CargoProcess"
                            }
                        }
                    }
                }
            }
        },
        "/image": {
            "post": {
                "description": "上传图片",
                "summary": "上传图片",
                "parameters": [
                    {
                        "type": "file",
                        "description": "文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status 200 表示成功 否则提示msg内容",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "msg": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "dto.CargoKindWithAttrs": {
            "type": "object",
            "properties": {
                "cargo_attrs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.CargoAttr"
                    }
                },
                "cargo_kind": {
                    "$ref": "#/definitions/model.CargoKind"
                }
            }
        },
        "dto.ReqAddCargo": {
            "type": "object",
            "properties": {
                "cargo": {
                    "$ref": "#/definitions/model.Cargo"
                },
                "cargo_attr_values": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.CargoAttrValue"
                    }
                },
                "measures": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Measure"
                    }
                }
            }
        },
        "model.Cargo": {
            "type": "object",
            "properties": {
                "cargo_code": {
                    "description": "货品编码",
                    "type": "string"
                },
                "cargo_id": {
                    "description": "货品ID",
                    "type": "integer"
                },
                "cargo_name": {
                    "description": "货品名称",
                    "type": "string"
                },
                "ck_id": {
                    "description": "货品种类ID",
                    "type": "integer"
                },
                "created_at": {
                    "type": "integer"
                },
                "image_name": {
                    "description": "大图",
                    "type": "string"
                },
                "status": {
                    "description": "状态 1 正常 8 删除",
                    "type": "integer"
                },
                "thumbnail_name": {
                    "description": "缩略图",
                    "type": "string"
                },
                "updated_at": {
                    "type": "integer"
                }
            }
        },
...
        "model.CargoKind": {
            "type": "object",
            "properties": {
                "ck_code": {
                    "description": "货品编码",
                    "type": "string"
                },
                "ck_id": {
                    "type": "integer"
                },
                "ck_name": {
                    "description": "货品名称",
                    "type": "string"
                },
                "created_at": {
                    "type": "integer"
                },
                "intro": {
                    "description": "货品简介",
                    "type": "string"
                },
                "status": {
                    "description": "状态 1 正常 8 删除",
                    "type": "integer"
                },
                "type": {
                    "description": "1物料 2制品",
                    "type": "integer"
                },
                "updated_at": {
                    "type": "integer"
                }
            }
        },
        "model.CargoProcess": {
            "type": "object",
            "properties": {
                "cargo_id": {
                    "description": "关联货品 cargo",
                    "type": "integer"
                },
                "created_at": {
                    "type": "integer"
                },
                "is_exact": {
                    "description": "是否精准",
                    "type": "boolean"
                },
                "max_use": {
                    "description": "当用量是精准时， 取最大值， 否则为区间",
                    "type": "number"
                },
                "measure_id": {
                    "description": "计量单位 关联 measure",
                    "type": "integer"
                },
                "min_use": {
                    "description": "当用量是精准时， 取最大值， 否则为区间",
                    "type": "number"
                },
                "order": {
                    "description": "生产顺序",
                    "type": "integer"
                },
                "process_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "integer"
                }
            }
        },
...
    }
}